pyproject.toml
[tool.poetry]
name = "chatbot-code-interpreter"
version = "0.1.0"
description = ""
authors = ["Dylan Castillo <dylanjcastillo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chatbot_code_interpreter"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.166"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
litestar = {extras = ["standard"], version = "^2.0.0a6"}


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
